requirements.txt
streamlit
openai
requests
beautifulsoup4
pdfplumber
python-pptx
python-docx
openpyxl